Implement LOAD/STORE pcodeops for simple stack operations
This is quite hacky. Notably, we are just storing locals
in one big 0x2000 byte alloca currently. Any functions that
want a stack frame bigger than this may cause a crash.

Also, we are assuming the MIPS stack pointer register is
zero so we can get the offset. Obviously this won't work
with other architectures.