Fix segfault on exit
This was caused by the me using the auto keyword, which made
it unclear that I was copying a BlockGraph object instead of
taking a reference to it.

Additionally, I've decided to let Ghidra handle destructing
and freeing the Funcdata objects instead of doing it with
a std::unique_ptr.